fix: quote image gallery
